requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
distlib==0.3.6
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
filelock==3.8.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
virtualenv==20.16.5
